+++ glamod_marine_processing/obs_suite/scripts/level1d.py
@@ -62,12 +62,12 @@ from collections import Counter
 from importlib import reload
 
 import pandas as pd
-import simplejson
 from _utilities import (
     FFS,
     date_handler,
     delimiter,
     paths_exist,
+    save_quicklook,
     script_setup,
     table_to_csv,
 )
@@ -107,9 +107,9 @@ def process_table(table_df, table_name):
             table_df.index
         ]
 
-    meta_dict[table_name] = {"total": len(table_df), "updated": 0}
+    ql_dict[table_name] = {"total": len(table_df), "updated": 0}
     if merge:
-        meta_dict[table_name] = {"total": len(table_df), "updated": 0}
+        ql_dict[table_name] = {"total": len(table_df), "updated": 0}
         table_df.set_index("primary_station_id", drop=False, inplace=True)
 
         if table_name == "header":
@@ -132,12 +132,12 @@ def process_table(table_df, table_name):
         table_df.update(meta_table[~meta_table.index.duplicated()])
 
         updated_locs = [x for x in table_df.index if x in meta_table.index]
-        meta_dict[table_name]["updated"] = len(updated_locs)
+        ql_dict[table_name]["updated"] = len(updated_locs)
 
         if table_name == "header":
             missing_ids = [x for x in table_df.index if x not in meta_table.index]
             if len(missing_ids) > 0:
-                meta_dict["non " + params.md_model + " ids"] = {
+                ql_dict["non " + params.md_model + " ids"] = {
                     k: v for k, v in Counter(missing_ids).items()
                 }
             history_add = ";{}. {}".format(history_tstmp, "metadata fix")
@@ -204,7 +204,7 @@ else:
     logging.info(f"Metadata not available for data source-deck {params.sid_dck}")
     logging.info("level1d data will be created with no merging")
 
-meta_dict = {}
+ql_dict = {}
 
 # DO THE DATA PROCESSING ------------------------------------------------------
 # -----------------------------------------------------------------------------
@@ -266,14 +266,6 @@ for table in obs_tables:
 
 # 4. SAVE QUICKLOOK -----------------------------------------------------------
 logging.info("Saving json quicklook")
-level_io_filename = os.path.join(params.level_ql_path, params.fileID + ".json")
-with open(level_io_filename, "w") as fileObj:
-    simplejson.dump(
-        {"-".join([params.year, params.month]): meta_dict},
-        fileObj,
-        default=date_handler,
-        indent=4,
-        ignore_nan=True,
-    )
-
-logging.info("End")
+save_quicklook(
+    params.level_qc_path, params.fileID, params.fileID_date, ql_dict, date_handler
+)