+++ CHANGES.rst
@@ -17,6 +17,7 @@ Internal changes
 * ``obs_suite``: update ICOADS_R3.0.0T release_4.0 time period list (:pull:`73`)
 * ``obs_suite``: combine loops over decks in ``level_slurm.py`` and ``config_array.py`` to one single loop (:issue:`74`, :pull:`79`)
 * ``obs_suie``: name log files according to the date of the source files instead of simply numbering them consecutively (:issue:`74`, :pull:`79`)
+* ``obs_suite``: standardize level scripts (:pull:`79`)
 
 v7.1.0 (2024-11-25)
 -------------------